Add updateCurrentBOSS
