+++ src/client/convex/updateBOSS.ts
@@ -1,0 +1,5 @@
+import * as Discord from 'discord.js'
+
+export const UpdateCurrentBOSS = async (msg: Discord.Message) => {
+  msg
+}
